feat(cli): add glob flags to cli tool
